金额问题

+++ tdx-ext/tdx-ext-vo.go
@@ -5,8 +5,8 @@ import "github.com/nie312122330/niexq-tdx/tdx"
 // 两家
 type TdxExtTodayMoney struct {
 	tdx.TdxFshqVo
-	PreClosePrice int `json:"preClosePrice"`
-	BigInMoney    int `json:"bigInMoney"`
-	BigOutMoney   int `json:"bigOutMoney"`
-	BigMoneyCount int `json:"bigMoneyCount"`
+	PreClosePrice int   `json:"preClosePrice"`
+	BigInMoney    int64 `json:"bigInMoney"`
+	BigOutMoney   int64 `json:"bigOutMoney"`
+	BigMoneyCount int64 `json:"bigMoneyCount"`
 }
